==> 03 Scripts/08_Task 4.6 Combining and Exporting Data.ipynb <==
# Contents
#### Importing libraries
#### Importing orders_products_combined
#### Combining orders_products_combined df with products df

# 01. Importing libraries

import pandas as pd
import numpy as np
import os

# 02. Importing data set

# Importing orders_products_combined dataframe from the pickle file I saved previously
df_orders_products_combined = pd.read_pickle(r'C:\Users\veren\Python Data\26-10-2021 Instacart Basket Analysis\02 Data\Prepared Data\orders_products_combined.pkl')

# Checking the shape of orders_products_combined (I used a subset of the data because it was too large for my computer)
df_orders_products_combined.shape

The check came out correct: The shape is the same that I expected for my subset.

# Determining a suitable way to combine df_orders_products_combined with products data set.
# Importing products data set
df_products = pd.read_csv(r'C:\Users\veren\Python Data\26-10-2021 Instacart Basket Analysis\02 Data\Prepared Data\products_cleaned.csv')

# Looking at df_products head to find the suitable column to merge
df_products.head()

# Removing the index column
df_products = df_products.drop(columns = ['Unnamed: 0',])

# Checking the head again
df_products.head()

# Checking the head of df_orders_products_combined to find the suitable column to merge
df_orders_products_combined.head()

product_id is a suitable column to merge the data sets.

# 03. Combining the dataframes

# Dropping the _merge column in df_orders_products_combined before the merge (not removing it will give out an error)
df_orders_products_combined_1 = df_orders_products_combined.drop(columns = ['_merge'])

# Merging
orders_products_merged = df_orders_products_combined_1.merge(df_products, on = ['product_id'], how = 'left',indicator = True)

# Checking at orders_products_merged with head()
orders_products_merged.head()

# Confirming the results of the merge using the merge flag
orders_products_merged['_merge'].value_counts()

orders_products_merged.shape

# Creating a path to export orders_products_merged as a pkl
path = r'C:\Users\veren\Python Data\26-10-2021 Instacart Basket Analysis'

path 

orders_products_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged.pkl'))
