==> 03 Scripts/11_Task 4.8 Grouping Data & Aggregating Variables.ipynb <==
### 01. Importing libraries

import pandas as pd
import numpy as np
import os

### 02. Importing dataframe

ords_prods_merge = pd.read_pickle(r'C:\Users\veren\Python Data\26-10-2021 Instacart Basket Analysis\02 Data\Prepared Data\ords_prods_merge_GDAV.pkl')

### 03. Task 4.8

# Find the aggregated mean of the “order_number” column grouped by “department_id” for the entire dataframe.
ords_prods_merge.groupby('department_id').agg({'order_number':['mean']})

The mean of the subset from the previous exercise and the mean of the whole dataset are very close. I think this makes sense because statistically the mean of the subset of the data can differ slightly from the whole, but it all should gravitate toward similar means.

# The marketing team at Instacart wants to know whether there's a difference between the spending habits of the three types of customers you identified. Use the loyalty flag you created and check the basic statistics of the product prices for each"
# What you're trying to determine is whether the prices of products purchased by loyal customers differ from those purchased by regular or new customers"
ords_prods_merge.groupby('loyalty_flag').agg({'prices': ['mean', 'min', 'max']})

New customers and regular customers tend to spend more in the 13 US-Dollar prize range while loyal customers purchase products that range around 10 US-Dollar.

# Create a spending flag for each user based on the average price across all their orders using the following criteria:
# If the mean of the prices of products purchased by a user is lower than 10, then flag them as a “Low spender"
# If the mean of the prices of products purchased by a user is higher than or equal to 10, then flag them as a “High spender.”
ords_prods_merge['spending'] = ords_prods_merge.groupby(['user_id'])['prices'].transform(np.mean)

ords_prods_merge.loc[ords_prods_merge['spending'] < 10, 'spending_flag'] = 'Low Spender'

ords_prods_merge.loc[ords_prods_merge['spending'] >= 10, 'spending_flag'] = 'High Spender'

# Pulling the data from the spending flag column
ords_prods_merge['spending_flag'].value_counts(dropna = False)

# Create an order frequency flag that marks the regularity of a user’s ordering behavior according to the median in the “days_since_prior_order” column. 
# The criteria for the flag should be as follows:
# If the median of “days_since_prior_order” is higher than 20, then the customer should be labeled a “Non-frequent customer"
# If the median is higher than 10 and lower than or equal to 20, then the customer should be labeled a “Regular customer.”
# If the median is lower than or equal to 10, then the customer should be labeled a “Frequent customer.”
ords_prods_merge['frequency'] = ords_prods_merge.groupby(['user_id'])['days_since_prior_order'].transform(np.median)

ords_prods_merge.loc[ords_prods_merge['frequency'] > 20, 'frequency_flag'] = 'Non-frequent customer'

ords_prods_merge.loc[(ords_prods_merge['frequency'] <= 20) & (ords_prods_merge['frequency'] > 10), 'frequency_flag'] = 'Regular customer'

ords_prods_merge.loc[ords_prods_merge['frequency'] <= 10, 'frequency_flag'] = 'Frequent customer'

# Pulling the data from the frequency flag column
ords_prods_merge['frequency_flag'].value_counts(dropna = False)

ords_prods_merge.head(10)

# Exporting file as pkl
path = r'C:\Users\veren\Python Data\26-10-2021 Instacart Basket Analysis'

path

ords_prods_merge.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_merge_GDAV_whole.pkl'))
