==> 03 Scripts/13_Task 4.9 Part 1 - Merging previous data with the new customers data.ipynb <==
### 01. Importing libraries

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

### 02. Importing customers dataset

# Creating the path
path = r'C:\Users\veren\Python Data\26-10-2021 Instacart Basket Analysis'

path

# Importing customers data set
df_customers = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'))

### 03. Checking the data quality and wrangling the data

# General checks
df_customers.shape

df_customers.head()

# Renaming the columns consistently
df_customers.rename(columns = {'First Name' : 'first_name', 'Surnam' : 'surname','Gender' : 'gender', 'STATE' : 'state', 'Age' : 'age', 'fam_status' : 'family_status'}, inplace = True)

# Check
df_customers.head(2)

I chose not to drop any columns.

# Using the describe function to "get to know" the customers data further
df_customers.describe()

# Looking for missing data
df_customers.isnull().sum()

I decided to not do anything about the missing first_name values because as long as I have the surname values the data is important and unique.

# Checking for columns with mixed data types in the customers dataset to prevent errors

for col in df_customers.columns.tolist():
  weird = (df_customers[[col]].applymap(type) != df_customers[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_customers[weird]) > 0:
    print (col)


df_customers['first_name'].dtype

# Converting the entries of the first_name column into a string
df_customers['first_name']= df_customers['first_name'].astype('str')

### 04. Combining the previous Instacart data with the new customers data

# Importing the most recent instacart data
df_ords_prods_merge = pd.read_pickle(r'C:\Users\veren\Python Data\26-10-2021 Instacart Basket Analysis\02 Data\Prepared Data\ords_prods_merge_GDAV_whole.pkl')

df_ords_prods_merge.shape

df_ords_prods_merge.head()

# Removing unnecessary data ("eval_set")

df_ords_prods_merge.drop(columns = ['eval_set'], inplace = True)

# Check
df_ords_prods_merge.columns

# Removing _merge column
df_ords_prods_merge.drop(columns = ['_merge'], inplace = True)

#Check 
df_ords_prods_merge.columns

# Renaming "busiest day" and "busiest days" for consistency
df_ords_prods_merge.rename(columns = {'busiest days' : 'busiest_days', 'busiest day': 'busiest_day'}, inplace = True)

# Check
df_ords_prods_merge.columns

# Checking the datatypes of df_ords_prods_merge to make sure the user_id column has the same data type in both data sets
df_ords_prods_merge.dtypes

df_customers.dtypes

The columns "user_id" is int64 in both data sets so merging both data frames on this column should work.

### 05. Merging the two data sets

# Combining the customer data with the previous Instacart data
df_ords_prods_cust_merged = df_ords_prods_merge.merge(df_customers, on = 'user_id')

df_ords_prods_cust_merged.shape

df_ords_prods_cust_merged.head(100)

# Exporting the data as pkl
path

df_ords_prods_cust_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_cust_merged_1.pkl'))
