==> 03 Scripts/16_Task 4.10 - Part 1 Data Security and Customer Regions.ipynb <==
### 01. Importing libraries

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

### 02. Importing current dataframe

ords_prods_cust = pd.read_pickle(r'C:\Users\veren\Python Data\26-10-2021 Instacart Basket Analysis\02 Data\Prepared Data\ords_prods_cust_merged_1.pkl')

ords_prods_cust.shape

ords_prods_cust.head()

ords_prods_cust.columns

ords_prods_cust.dtypes

### 03. Addressing Personally Identifiable Information (PII data)

# Addressing the PII problem by creating a new dataframe without the "first_name" column and the "surname" column
ords_prods_cust_checked = ords_prods_cust.drop(columns = ['first_name', 'surname'])

ords_prods_cust_checked.columns

ords_prods_cust_checked.shape

### 04. Customer Regions

# Creating a flag for the Northeast region based on the article in the task description
ords_prods_cust_checked.loc[ords_prods_cust_checked['state'].isin(['Maine', 'New Hampshire', 'Vermont', 'Massachusetts', 'Rhode Island', 'Connecticut', 'New York', 'Pennsylvania', 'New Jersey']), 'region'] = 'Northeast'

# Creating a flag for the Midwest region
ords_prods_cust_checked.loc[ords_prods_cust_checked['state'].isin(['Wisconsin', 'Michigan', 'Illinois', 'Indiana', 'Ohio', 'North Dakota', 'South Dakota', 'Nebraska', 'Kansas', 'Minnesota', 'Iowa', 'Missouri']), 'region'] = 'Midwest'

# Creating a flag for the South region
ords_prods_cust_checked.loc[ords_prods_cust_checked['state'].isin(['Delaware', 'Maryland', 'District of Columbia', 'Virginia', 'West Virginia', 'North Carolina', 'South Carolina', 'Georgia', 'Florida', 'Kentucky', 'Tennessee', 'Mississippi', 'Alabama', 'Oklahoma', 'Texas', 'Arkansas', 'Louisiana']), 'region'] = 'South'

# Creating a flag for the West region
ords_prods_cust_checked.loc[ords_prods_cust_checked['state'].isin(['Idaho', 'Montana', 'Wyoming', 'Nevada', 'Utah', 'Colorado', 'Arizona', 'New Mexico', 'Alaska', 'Washington', 'Oregon', 'California', 'Hawaii']), 'region'] = 'West'

ords_prods_cust_checked['region'].value_counts(dropna = False)

# Exporting ords_prods_cust_checked
path = r'C:\Users\veren\Python Data\26-10-2021 Instacart Basket Analysis'

ords_prods_cust_checked.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_cust_region.pkl'))
