==> 03 Scripts/17_Task 4.10 - Part 2 Regional Spending and Low-activity Customers.ipynb <==
### 01. Importing libraries

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

### 02. Importing data frame

ords_prods_cust = pd.read_pickle(r'C:\Users\veren\Python Data\26-10-2021 Instacart Basket Analysis\02 Data\Prepared Data\ords_prods_cust_region.pkl')

ords_prods_cust.shape

ords_prods_cust.head()

ords_prods_cust.columns

### 03. Comparing regional spending

# Crossing the region variable with the spending_flag
crosstab = pd.crosstab(ords_prods_cust['region'], ords_prods_cust['spending_flag'], dropna = False)

crosstab

The South region has the most high spenders, followed by the West, the Midwest, and the Northeast. But, the South also has the most low spenders, followed by the West, the Midwest, and the Northeast. Overall, these numbers seem to be correlated to the population counts in these regions because the most populated region is the South, followed by the West, the Midwest, and the Northeast.

crosstab.to_clipboard

### 04. Creating an exclusion flag for low-activity customers

# Creating a flag for low-activity customers (who have less than 5 orders)
ords_prods_cust.loc[ords_prods_cust['max_order'] < 5, 'exclusion_flag'] = 'Low-activity Customer'

# Creating a flag for customers with a higher activity (who have 5 or more orders). I call them non_low_activity customer.
ords_prods_cust.loc[ords_prods_cust['max_order'] >= 5, 'exclusion_flag'] = 'Non-low-activity Customer'

# Checking the creation of the flag with the head()-function
ords_prods_cust[['max_order', 'exclusion_flag']].head(10)

ords_prods_cust['exclusion_flag'].value_counts(dropna = False)

### 05. Creating a dataframe for low-activity customers and non-low-activity customers

# Creating a new dataframe only showing data for low-activity customers
df_low_cust = ords_prods_cust[ords_prods_cust['exclusion_flag'] == 'Low-activity Customer']

df_low_cust.head()

df_low_cust.shape

# Creating a new dataframe only showing data for non-low-activity customers
df_normal_cust = ords_prods_cust[ords_prods_cust['exclusion_flag'] == 'Non-low-activity Customer']

df_normal_cust.head()

df_normal_cust.shape

# Checking if df_normal_cust and df_low_cust together have the same number of rows than ords_prods_cust
29329778 + 1026350

Success!

### 06. Exporting dataframes

path = r'C:\Users\veren\Python Data\26-10-2021 Instacart Basket Analysis'

path

# Exporting df_normal_cust
df_normal_cust.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'normal_customers.pkl'))

# Exporting ords_prods_cust
ords_prods_cust.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'final_ords_prods_cust_whole.pkl'))
