==> 03 Scripts/18_Task 4.10 - Part 3 Customer Profiling - Age Groups and Income Ranges.ipynb <==
### 01. Importing libraries

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

### 02. Importing data frame

ords_prods_cust = pd.read_pickle(r'C:\Users\veren\Python Data\26-10-2021 Instacart Basket Analysis\02 Data\Prepared Data\normal_customers.pkl')

# Checking the columns and records for customer profiling options
pd.options.display.max_columns = None
ords_prods_cust.head()

### 03. Defining age groups

# Defining a range of age groups (based on pewresearch.org)
# 18-29, 30-49, 50-64, 65+
# Defining age group 18-29
ords_prods_cust.loc[(ords_prods_cust['age'] >= 18) & (ords_prods_cust['age'] <= 29), 'age_group'] = '18-29'

# Defining age group 30-49
ords_prods_cust.loc[(ords_prods_cust['age'] >= 30) & (ords_prods_cust['age'] <= 49), 'age_group'] = '30-49'

# Defining age group 50-64
ords_prods_cust.loc[(ords_prods_cust['age'] >= 50) & (ords_prods_cust['age'] <= 64), 'age_group'] = '50-64'

# Defining age group 65+
ords_prods_cust.loc[ords_prods_cust['age'] >= 65, 'age_group'] = '65+'

ords_prods_cust['age_group'].value_counts(dropna = False)

### 04. Defining income ranges

ords_prods_cust['income'].value_counts().sort_index()

# Defining the income_range
# range: 25.903 - 593.901
# 25.000 - 50.000
# 50.001 - 100.000
# 100.001 - 150.000
# 150.001 - 200.000
# 200.000+
# Defining the first income range 25.000 - 50.000
ords_prods_cust.loc[(ords_prods_cust['income'] >= 25000) & (ords_prods_cust['income'] <= 50000), 'income_range'] = '25.000-50.000'

# Defining the second income range 50.001 - 100.000
ords_prods_cust.loc[(ords_prods_cust['income'] >= 50001) & (ords_prods_cust['income'] <= 100000), 'income_range'] = '50.000-100.000'

# Defining the third income range 100.001 - 150.000
ords_prods_cust.loc[(ords_prods_cust['income'] >= 100001) & (ords_prods_cust['income'] <= 150000), 'income_range'] = '100.000-150.000'

# Defining the fourth income range 150.001 - 200.000
ords_prods_cust.loc[(ords_prods_cust['income'] >= 150001) & (ords_prods_cust['income'] <= 200000), 'income_range'] = '150.000-200.000'

# Defining the fourth income range 200.000+
ords_prods_cust.loc[ords_prods_cust['income'] > 200000, 'income_range'] = '200.000+'

ords_prods_cust['income_range'].value_counts(dropna = False)

pd.options.display.max_columns = None
ords_prods_cust.head()

# Crossing income_range with age_group
crosstab_age_income = pd.crosstab(ords_prods_cust['age_group'], ords_prods_cust['income_range'], dropna = False).sort_index()

crosstab_age_income

# Exporting dataframe
# Creating a path
path = r'C:\Users\veren\Python Data\26-10-2021 Instacart Basket Analysis'

path

ords_prods_cust.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'final_income_age.pkl'))
