==> 03 Scripts/19_Task 4.10 - Part 4 Customer Profiling - Goods and Days of the Week.ipynb <==
### 01. Importing libraries

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

### 02. Importing dataframe

df_final = pd.read_pickle(r'C:\Users\veren\Python Data\26-10-2021 Instacart Basket Analysis\02 Data\Prepared Data\final_income_age.pkl')

pd.options.display.max_columns = None
df_final.head()

### 03. Defining department names

# Adding the name of the department based on the code in an earlier data dictionary
# Adding the frozen department
df_final.loc[df_final['department_id'] == 1, 'department'] = 'frozen'

# adding the other department
df_final.loc[df_final['department_id'] == 2, 'department'] = 'other'

# Adding the bakery department
df_final.loc[df_final['department_id'] == 3, 'department'] = 'bakery'

# Adding the produce department
df_final.loc[df_final['department_id'] == 4, 'department'] = 'produce'

# Adding the alcohol department
df_final.loc[df_final['department_id'] == 5, 'department'] = 'alcohol'

# Adding the international department
df_final.loc[df_final['department_id'] == 6, 'department'] = 'international'

# Adding the beverages department
df_final.loc[df_final['department_id'] == 7, 'department'] = 'beverages'

# Adding the pets department
df_final.loc[df_final['department_id'] == 8, 'department'] = 'pets'

# Adding the dry goods and pasta department
df_final.loc[df_final['department_id'] == 9, 'department'] = 'dry goods & pasta'

# Adding the bulk department
df_final.loc[df_final['department_id'] == 10, 'department'] = 'bulk'

# Adding the personal care department
df_final.loc[df_final['department_id'] == 11, 'department'] = 'personal care'

# Adding the meat and seafood department
df_final.loc[df_final['department_id'] == 12, 'department'] = 'meat & seafood'

# Adding the pantry department
df_final.loc[df_final['department_id'] == 13, 'department'] = 'pantry'

# Adding the breakfast department
df_final.loc[df_final['department_id'] == 14, 'department'] = 'breakfast'

# Adding the canned goods department
df_final.loc[df_final['department_id'] == 15, 'department'] = 'canned goods'

# Adding the dairy and eggs department
df_final.loc[df_final['department_id'] == 16, 'department'] = 'dairy & eggs'

# Adding the household department
df_final.loc[df_final['department_id'] == 17, 'department'] = 'household'

# Adding the babies department
df_final.loc[df_final['department_id'] == 18, 'department'] = 'babies'

# Adding the snacks department
df_final.loc[df_final['department_id'] == 19, 'department'] = 'snacks'

# Adding the deli department
df_final.loc[df_final['department_id'] == 20, 'department'] = 'deli'

# Adding the missing department
df_final.loc[df_final['department_id'] == 21, 'department'] = 'missing'

# Counting orders by departments
df_final['department'].value_counts()

### 04. Defining proper days of the week

# Adding Saturday
df_final.loc[df_final['order_dow'] == 0, 'day_of_week'] = 'Saturday'

# Adding Sunday
df_final.loc[df_final['order_dow'] == 1, 'day_of_week'] = 'Sunday'

# Adding Monday
df_final.loc[df_final['order_dow'] == 2, 'day_of_week'] = 'Monday'

# Adding Tuesday
df_final.loc[df_final['order_dow'] == 3, 'day_of_week'] = 'Tuesday'

# Adding Wednesday
df_final.loc[df_final['order_dow'] == 4, 'day_of_week'] = 'Wednesday'

# Adding Thursday
df_final.loc[df_final['order_dow'] == 5, 'day_of_week'] = 'Thursday'

# Adding Friday
df_final.loc[df_final['order_dow'] == 6, 'day_of_week'] = 'Friday'

# Counting orders by day of the week
df_final['day_of_week'].value_counts()

pd.options.display.max_columns = None
df_final.head()

### 05. Producing crosstabs for different indicators like family status, departments or region

# Counting orders by family status
df_final['family_status'].value_counts()

# Crossing orders by department and family status
crosstab_family_status_dept = pd.crosstab(df_final['department'], df_final['family_status'], dropna = False)

crosstab_family_status_dept

# Crossing orders by region and family status
crosstab_family_status_region = pd.crosstab(df_final['family_status'], df_final['region'], dropna = False)

crosstab_family_status_region

# Crossing age_range and family status
crosstab_family_status_age = pd.crosstab(df_final['family_status'], df_final['age_group'], dropna = False)

crosstab_family_status_age 

### 06. Exporting dataframe

# Creating a path
path = r'C:\Users\veren\Python Data\26-10-2021 Instacart Basket Analysis'

path

# Exporting df_final
df_final.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'final_complete.pkl'))
